[bert_squeeze/utils/scorers] - refactor: restructure scorers for sequence classification
 - Rename scorer.py to sequence_classification_scorer.py
 - Introduce BaseSequenceClassificationScorer class
 - Update LooseScorer to inherit from BaseSequenceClassificationScorer
 - Add FastBertSequenceClassificationScorer for FastBert models
 - Remove unused scorer_fast.py
