+++ bert_squeeze/utils/scorers/__init__.py
@@ -1,2 +1,5 @@
-from .scorer import LooseScorer, Scorer
-from .scorer_fast import FastBertScorer
+from .sequence_classification_scorer import (
+    BaseSequenceClassificationScorer,
+    FastBertSequenceClassificationScorer,
+    LooseSequenceClassificationScorer,
+)
